Add GitHub Action + fix Linux build (#2)
# Description

Adds a GitHub Action for automated testing.

Also fixes building on Linux, which I think just wasn't working at all.
The issues were:
- not being able to find `assert.h` because I was missing a `linkLibC()`
on the QBE exe
- the creation of the `config.h` being wrong insofar as it was being
generated into the directory where `zig build` is run, not into the
`src/` directory as expected. To make this work properly I am now using
`addWriteFiles` (see
https://ziglang.org/learn/build-system/#write-files) to generate the
`config.h` into `zig-cache` and then making sure to `addIncludePath`
with `wf.getDirectory()` so that `#include "config.h"` finds the file in
the cache (instead of in `src/). I believe this is the way.