Update upstream qbe version + rearrange repo
This updates to the latest master qbe. The nice thing that has happened
since the initial version pulled in to this repo is that they have fixed
the `asm` field name that was causing issues when compiling with zig (it
is now `fmt`).

As a consequence we no longer have to manually modify the original source
to allow building with zig. This gets us very close to not having to pull
in the upstream repo at all (where we could instead fetch it as part of
build.zig.zon). However, the zig build system does seem able to depend
on cgit repos at the moment and the current release tarball is from about
a year ago...therefore we still for the moment need to pull the code in.

Based on previous releases we might get a new version in February, at which
point this repo can be mostly empty. Additionally, if the zig build system
is updated to support cgit, or indeed cgit is updated so that it is compatible
with zig, then we can have a version of this repo that keeps up with the
upstream master without manually pulling in the code.